backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
pandas==2.1.3
numpy==1.26.2
psycopg2-binary
pytz==2023.3.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
prometheus-client==0.19.0
python-json-logger==2.0.7
